refactor(feat/#25): 폴더명 변경 및 컴포넌트 분리
